// src/models/RevisionBucket.ts
import mongoose, { Schema, Document } from "mongoose";

export interface IRevisionBucket extends Document {
  userId: mongoose.Types.ObjectId;
  bucketType: string;
  questions: mongoose.Types.ObjectId[];
  autoGeneratedAt: Date;
  priorityLevel: number;
}

const revisionBucketSchema = new Schema<IRevisionBucket>(
  {
    userId: { type: Schema.Types.ObjectId, ref: "User", required: true, index: true },
    bucketType: { type: String, required: true, index: true },
    questions: { type: [Schema.Types.ObjectId], ref: "Question", default: [] },
    autoGeneratedAt: { type: Date, default: Date.now },
    priorityLevel: { type: Number, default: 1 },
  },
  { timestamps: true },
);

revisionBucketSchema.index({ userId: 1, bucketType: 1 });

export const RevisionBucket = mongoose.model<IRevisionBucket>("RevisionBucket", revisionBucketSchema);